requirements.txt
langchain>=0.3.25
langchain-community
langchain-ollama
langchain-postgres>=0.0.14
python-dotenv>=1.1.0
pypdf>=5.6.0
streamlit>=1.32.0
